Use BigInt in fibonacci stream

+++ src/main/scala/com/chriswk/common/MathHelpers.scala
@@ -11,11 +11,11 @@ object MathHelpers {
       case None => List(num)
     }
   }
-  lazy val fib: Stream[Int] = 0 #:: 1 #:: fib.zip(fib.tail).map(p => p._1 + p._2)
+  lazy val fib: Stream[BigInt] = 0 #:: 1 #:: fib.zip(fib.tail).map(p => p._1 + p._2)
   def gcd(a: BigInt, b: BigInt): BigInt = if (b == 0) a else gcd(b, a%b)
   def lcm(a: BigInt, b: BigInt): BigInt = a*b / gcd(a,b)
-  def even(a: Int) = a % 2 == 0
-  def odd(a: Int) = a % 2 == 1
+  def even(a: BigInt) = a % 2 == 0
+  def odd(a: BigInt) = a % 2 == 1
   def square(a: Int) = a * a
   def sumOfSquares(start: Int, finish: Int): BigInt = (start to finish).map(square).sum
   def squareOfSum(start: Int, finish: Int): BigInt = square((start to finish).sum)
